+++ CoreArrangementPlugins/src/au/gov/asd/tac/constellation/plugins/arrangements/proximity/FR2DArranger.java
@@ -23,8 +23,8 @@ import au.gov.asd.tac.constellation.plugins.PluginInteraction;
 import au.gov.asd.tac.constellation.plugins.arrangements.Arranger;
 import au.gov.asd.tac.constellation.plugins.arrangements.GraphUtilities;
 import java.awt.geom.Point2D;
+import java.security.SecureRandom;
 import java.util.ArrayList;
-import java.util.Random;
 
 /**
  * main module to arrange a graph using the FR2D algorithm
@@ -57,7 +57,7 @@ class FR2DArranger implements Arranger {
 
     private final PluginInteraction interaction;
     
-    private final Random r = new Random();
+    private final SecureRandom r = new SecureRandom();
 
     /**
      *