+++ CoreTesting/src/au/gov/asd/tac/constellation/testing/construction/SphereGraphBuilderPlugin.java
@@ -45,12 +45,12 @@ import au.gov.asd.tac.constellation.utilities.font.FontUtilities;
 import au.gov.asd.tac.constellation.utilities.icon.IconManager;
 import au.gov.asd.tac.constellation.utilities.visual.LineStyle;
 import java.awt.Font;
+import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
-import java.util.Random;
 import java.util.stream.Collectors;
 import org.openide.util.Exceptions;
 import org.openide.util.NbBundle.Messages;
@@ -81,12 +81,12 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
     public static final String USE_ALL_DISPLAYABLE_CHARS_PARAMETER_ID = PluginParameter.buildId(SphereGraphBuilderPlugin.class, "use_all_displayable_chars");
     public static final String DRAW_MANY_TX_PARAMETER_ID = PluginParameter.buildId(SphereGraphBuilderPlugin.class, "draw_many_tx");
     public static final String DRAW_MANY_DECORATORS_PARAMETER_ID = PluginParameter.buildId(SphereGraphBuilderPlugin.class, "draw_many_deco");
-
-    private final Random random = new Random();
-    
+   
     private static final String NODE = "~Node ";
     private static final String TYPE = "~Type ";
     private static final String BACKGROUND_ROUND_CIRCLE = "Background.Round Circle";
+    
+    private final SecureRandom random = new SecureRandom();
 
     @Override
     public String getDescription() {
@@ -338,7 +338,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
             graph.setBooleanValue(vxIsGoodAttr, vxId, vx % 2 == 0);
             graph.setStringValue(vxCountry1Attr, vxId, countries.get(vx % countries.size()));
             graph.setStringValue(vxCountry2Attr, vxId, countries.get((vx + 1) % countries.size()));
-            graph.setFloatValue(vxNormalisedAttr, vxId, (float) Math.random());
+            graph.setFloatValue(vxNormalisedAttr, vxId, random.nextFloat());
 
             vxIds[vx] = vxId;
             vx++;
@@ -468,7 +468,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
         graph.setStringValue(vxBackgroundIconAttr, vx0, BACKGROUND_ROUND_CIRCLE);
         graph.setStringValue(vxCountry1Attr, vx0, countries.get(vx0 % countries.size()));
         graph.setStringValue(vxCountry2Attr, vx0, countries.get((vx0 + 1) % countries.size()));
-        graph.setFloatValue(vxNormalisedAttr, vx0, (float) Math.random());
+        graph.setFloatValue(vxNormalisedAttr, vx0, random.nextFloat());
         graph.setObjectValue(attrBlaze, vx0, new Blaze(random.nextInt(360), randomColor3(random)));
 
         final int vx1 = graph.addVertex();
@@ -484,7 +484,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
         graph.setStringValue(vxBackgroundIconAttr, vx1, BACKGROUND_ROUND_CIRCLE);
         graph.setStringValue(vxCountry1Attr, vx1, countries.get(vx1 % countries.size()));
         graph.setStringValue(vxCountry2Attr, vx1, countries.get((vx1 + 1) % countries.size()));
-        graph.setFloatValue(vxNormalisedAttr, vx1, (float) Math.random());
+        graph.setFloatValue(vxNormalisedAttr, vx1, random.nextFloat());
         graph.setObjectValue(attrBlaze, vx1, new Blaze(random.nextInt(360), randomColor3(random)));
 
         final int vx2 = graph.addVertex();
@@ -500,7 +500,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
         graph.setStringValue(vxBackgroundIconAttr, vx2, BACKGROUND_ROUND_CIRCLE);
         graph.setStringValue(vxCountry1Attr, vx2, countries.get(vx2 % countries.size()));
         graph.setStringValue(vxCountry2Attr, vx2, countries.get((vx2 + 1) % countries.size()));
-        graph.setFloatValue(vxNormalisedAttr, vx2, (float) Math.random());
+        graph.setFloatValue(vxNormalisedAttr, vx2, random.nextFloat());
         graph.setObjectValue(attrBlaze, vx2, new Blaze(random.nextInt(360), randomColor3(random)));
 
         final int vx3 = graph.addVertex();
@@ -516,7 +516,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
         graph.setStringValue(vxBackgroundIconAttr, vx2, BACKGROUND_ROUND_CIRCLE);
         graph.setStringValue(vxCountry1Attr, vx3, countries.get(vx3 % countries.size()));
         graph.setStringValue(vxCountry2Attr, vx3, countries.get((vx3 + 1) % countries.size()));
-        graph.setFloatValue(vxNormalisedAttr, vx3, (float) Math.random());
+        graph.setFloatValue(vxNormalisedAttr, vx3, random.nextFloat());
         graph.setObjectValue(attrBlaze, vx3, new Blaze(random.nextInt(360), randomColor3(random)));
 
         final int vx4 = graph.addVertex();
@@ -532,7 +532,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
         graph.setStringValue(vxBackgroundIconAttr, vx2, BACKGROUND_ROUND_CIRCLE);
         graph.setStringValue(vxCountry1Attr, vx4, countries.get(vx4 % countries.size()));
         graph.setStringValue(vxCountry2Attr, vx4, countries.get((vx4 + 1) % countries.size()));
-        graph.setFloatValue(vxNormalisedAttr, vx4, (float) Math.random());
+        graph.setFloatValue(vxNormalisedAttr, vx4, random.nextFloat());
         graph.setObjectValue(attrBlaze, vx4, new Blaze(random.nextInt(360), randomColor3(random)));
 
         final int vx5 = graph.addVertex();
@@ -548,7 +548,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
         graph.setStringValue(vxBackgroundIconAttr, vx2, BACKGROUND_ROUND_CIRCLE);
         graph.setStringValue(vxCountry1Attr, vx5, countries.get(vx5 % countries.size()));
         graph.setStringValue(vxCountry2Attr, vx5, countries.get((vx5 + 1) % countries.size()));
-        graph.setFloatValue(vxNormalisedAttr, vx5, (float) Math.random());
+        graph.setFloatValue(vxNormalisedAttr, vx5, random.nextFloat());
         graph.setObjectValue(attrBlaze, vx5, new Blaze(random.nextInt(360), randomColor3(random)));
 
         // Draw multiple lines with offsets between two fixed nodes.
@@ -659,7 +659,7 @@ public class SphereGraphBuilderPlugin extends SimpleEditPlugin {
      *
      * @return A random RGB color.
      */
-    private static ConstellationColor randomColor3(Random r) {
+    private static ConstellationColor randomColor3(SecureRandom r) {
         return ConstellationColor.getColorValue(r.nextFloat(), r.nextFloat(), r.nextFloat(), 1f);
     }
 }